VST caculation
This code is used to calculate the VST (vertical shift transformation). For the model, the conversion coefficient beta can be directly calculated for two separate periods of time, and the result after lifting can be obtained. For the time continuity of the observations, the beta is calculated from the trend, and the result is multiplied by the number of years to amplify the corresponding change.